Module google_recommender1_beta1::api

Structs

BillingAccountLocationInsightTypeInsightGetCall
Gets the requested insight. Requires the recommender.*.get IAM permission for the specified insight type.

BillingAccountLocationInsightTypeInsightListCall
Lists insights for a Cloud project. Requires the recommender.*.list IAM permission for the specified insight type.

BillingAccountLocationInsightTypeInsightMarkAcceptedCall
Marks the Insight State as Accepted. Users can use this method to indicate to the Recommender API that they have applied some action based on the insight. This stops the insight content from being updated. MarkInsightAccepted can be applied to insights in ACTIVE state. Requires the recommender.*.update IAM permission for the specified insight.

BillingAccountLocationRecommenderRecommendationGetCall
Gets the requested recommendation. Requires the recommender.*.get IAM permission for the specified recommender.

BillingAccountLocationRecommenderRecommendationListCall
Lists recommendations for a Cloud project. Requires the recommender.*.list IAM permission for the specified recommender.

BillingAccountLocationRecommenderRecommendationMarkClaimedCall
Marks the Recommendation State as Claimed. Users can use this method to indicate to the Recommender API that they are starting to apply the recommendation themselves. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationClaimed can be applied to recommendations in CLAIMED or ACTIVE state. Requires the recommender.*.update IAM permission for the specified recommender.

BillingAccountLocationRecommenderRecommendationMarkFailedCall
Marks the Recommendation State as Failed. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation failed. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationFailed can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

BillingAccountLocationRecommenderRecommendationMarkSucceededCall
Marks the Recommendation State as Succeeded. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation was successful. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationSucceeded can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

BillingAccountMethods
A builder providing access to all methods supported on billingAccount resources. It is not used directly, but through the Recommender hub.

FolderLocationInsightTypeInsightGetCall
Gets the requested insight. Requires the recommender.*.get IAM permission for the specified insight type.

FolderLocationInsightTypeInsightListCall
Lists insights for a Cloud project. Requires the recommender.*.list IAM permission for the specified insight type.

FolderLocationInsightTypeInsightMarkAcceptedCall
Marks the Insight State as Accepted. Users can use this method to indicate to the Recommender API that they have applied some action based on the insight. This stops the insight content from being updated. MarkInsightAccepted can be applied to insights in ACTIVE state. Requires the recommender.*.update IAM permission for the specified insight.

FolderLocationRecommenderRecommendationGetCall
Gets the requested recommendation. Requires the recommender.*.get IAM permission for the specified recommender.

FolderLocationRecommenderRecommendationListCall
Lists recommendations for a Cloud project. Requires the recommender.*.list IAM permission for the specified recommender.

FolderLocationRecommenderRecommendationMarkClaimedCall
Marks the Recommendation State as Claimed. Users can use this method to indicate to the Recommender API that they are starting to apply the recommendation themselves. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationClaimed can be applied to recommendations in CLAIMED or ACTIVE state. Requires the recommender.*.update IAM permission for the specified recommender.

FolderLocationRecommenderRecommendationMarkFailedCall
Marks the Recommendation State as Failed. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation failed. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationFailed can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

FolderLocationRecommenderRecommendationMarkSucceededCall
Marks the Recommendation State as Succeeded. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation was successful. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationSucceeded can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

FolderMethods
A builder providing access to all methods supported on folder resources. It is not used directly, but through the Recommender hub.

GoogleCloudRecommenderV1beta1CostProjection
Contains metadata about how much money a recommendation can save or incur.

GoogleCloudRecommenderV1beta1Impact
Contains the impact a recommendation can have for a given category.

GoogleCloudRecommenderV1beta1Insight
An insight along with the information used to derive the insight. The insight may have associated recomendations as well.

GoogleCloudRecommenderV1beta1InsightRecommendationReference
Reference to an associated recommendation.

GoogleCloudRecommenderV1beta1InsightStateInfo
Information related to insight state.

GoogleCloudRecommenderV1beta1ListInsightsResponse
Response to the ListInsights method.

GoogleCloudRecommenderV1beta1ListRecommendationsResponse
Response to the ListRecommendations method.

GoogleCloudRecommenderV1beta1MarkInsightAcceptedRequest
Request for the MarkInsightAccepted method.

GoogleCloudRecommenderV1beta1MarkRecommendationClaimedRequest
Request for the MarkRecommendationClaimed Method.

GoogleCloudRecommenderV1beta1MarkRecommendationFailedRequest
Request for the MarkRecommendationFailed Method.

GoogleCloudRecommenderV1beta1MarkRecommendationSucceededRequest
Request for the MarkRecommendationSucceeded Method.

GoogleCloudRecommenderV1beta1Operation
Contains an operation for a resource loosely based on the JSON-PATCH format with support for: * Custom filters for describing partial array patch. * Extended path values for describing nested arrays. * Custom fields for describing the resource for which the operation is being described. * Allows extension to custom operations not natively supported by RFC6902. See https://tools.ietf.org/html/rfc6902 for details on the original RFC.

GoogleCloudRecommenderV1beta1OperationGroup
Group of operations that need to be performed atomically.

GoogleCloudRecommenderV1beta1Recommendation
A recommendation along with a suggested action. E.g., a rightsizing recommendation for an underutilized VM, IAM role recommendations, etc

GoogleCloudRecommenderV1beta1RecommendationContent
Contains what resources are changing and how they are changing.

GoogleCloudRecommenderV1beta1RecommendationInsightReference
Reference to an associated insight.

GoogleCloudRecommenderV1beta1RecommendationStateInfo
Information for state. Contains state and metadata.

GoogleCloudRecommenderV1beta1ValueMatcher
Contains various matching options for values for a GCP resource field.

GoogleTypeMoney
Represents an amount of money with its currency type.

OrganizationLocationInsightTypeInsightGetCall
Gets the requested insight. Requires the recommender.*.get IAM permission for the specified insight type.

OrganizationLocationInsightTypeInsightListCall
Lists insights for a Cloud project. Requires the recommender.*.list IAM permission for the specified insight type.

OrganizationLocationInsightTypeInsightMarkAcceptedCall
Marks the Insight State as Accepted. Users can use this method to indicate to the Recommender API that they have applied some action based on the insight. This stops the insight content from being updated. MarkInsightAccepted can be applied to insights in ACTIVE state. Requires the recommender.*.update IAM permission for the specified insight.

OrganizationLocationRecommenderRecommendationGetCall
Gets the requested recommendation. Requires the recommender.*.get IAM permission for the specified recommender.

OrganizationLocationRecommenderRecommendationListCall
Lists recommendations for a Cloud project. Requires the recommender.*.list IAM permission for the specified recommender.

OrganizationLocationRecommenderRecommendationMarkClaimedCall
Marks the Recommendation State as Claimed. Users can use this method to indicate to the Recommender API that they are starting to apply the recommendation themselves. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationClaimed can be applied to recommendations in CLAIMED or ACTIVE state. Requires the recommender.*.update IAM permission for the specified recommender.

OrganizationLocationRecommenderRecommendationMarkFailedCall
Marks the Recommendation State as Failed. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation failed. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationFailed can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

OrganizationLocationRecommenderRecommendationMarkSucceededCall
Marks the Recommendation State as Succeeded. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation was successful. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationSucceeded can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

OrganizationMethods
A builder providing access to all methods supported on organization resources. It is not used directly, but through the Recommender hub.

ProjectLocationInsightTypeInsightGetCall
Gets the requested insight. Requires the recommender.*.get IAM permission for the specified insight type.

ProjectLocationInsightTypeInsightListCall
Lists insights for a Cloud project. Requires the recommender.*.list IAM permission for the specified insight type.

ProjectLocationInsightTypeInsightMarkAcceptedCall
Marks the Insight State as Accepted. Users can use this method to indicate to the Recommender API that they have applied some action based on the insight. This stops the insight content from being updated. MarkInsightAccepted can be applied to insights in ACTIVE state. Requires the recommender.*.update IAM permission for the specified insight.

ProjectLocationRecommenderRecommendationGetCall
Gets the requested recommendation. Requires the recommender.*.get IAM permission for the specified recommender.

ProjectLocationRecommenderRecommendationListCall
Lists recommendations for a Cloud project. Requires the recommender.*.list IAM permission for the specified recommender.

ProjectLocationRecommenderRecommendationMarkClaimedCall
Marks the Recommendation State as Claimed. Users can use this method to indicate to the Recommender API that they are starting to apply the recommendation themselves. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationClaimed can be applied to recommendations in CLAIMED or ACTIVE state. Requires the recommender.*.update IAM permission for the specified recommender.

ProjectLocationRecommenderRecommendationMarkFailedCall
Marks the Recommendation State as Failed. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation failed. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationFailed can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

ProjectLocationRecommenderRecommendationMarkSucceededCall
Marks the Recommendation State as Succeeded. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation was successful. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state. MarkRecommendationSucceeded can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state. Requires the recommender.*.update IAM permission for the specified recommender.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Recommender hub.

Recommender
Central instance to access all Recommender related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.